Version 1.0

+++ src/main/java/fr/colline/monatis/comptes/service/CompteService.java
@@ -4,54 +4,65 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-import org.springframework.util.Assert;
 
-import fr.colline.monatis.comptes.CompteFonctionnelleErreur;
-import fr.colline.monatis.comptes.CompteTechniqueErreur;
 import fr.colline.monatis.comptes.model.Compte;
 import fr.colline.monatis.comptes.repository.CompteRepository;
 import fr.colline.monatis.exceptions.ServiceException;
+import fr.colline.monatis.exceptions.erreurs.ErreurFonctionnelle;
+import fr.colline.monatis.exceptions.erreurs.ErreurProgrammation;
+import fr.colline.monatis.exceptions.erreurs.ErreurTechnique;
+import jakarta.annotation.Nullable;
 
-@Service
 public abstract class CompteService<T extends Compte> {
-	
-	public T rechercherParId(Long id) throws ServiceException {
 
-		Assert.notNull(id, () -> "L'ID pour la recherche d'un compte de type '" + getTClass().getSimpleName() + "' est obligatoire");
+	public T rechercherParId(Long compteId) throws ServiceException {
+
+		if ( compteId == null ) {
+			throw new ServiceException(
+					ErreurProgrammation.ID_NULL,
+					getTClass().getSimpleName());
+		}
 
 		try {
-			Optional<T> optional = getRepository().findById(id);
+			Optional<T> optional = getRepository().findById(compteId);
 			return optional.isEmpty() ? null : optional.get();
 		}
 		catch (Throwable t) {
 			throw new ServiceException(
 					t,
-					CompteTechniqueErreur.RECHERCHE_PAR_ID,
-					getTClass().getSimpleName(),
-					id );
+					ErreurTechnique.TECH_RECHERCHE_COMPTE_PAR_ID,
+					compteId,
+					getTClass().getSimpleName() );
 		}
 	}
 
-	public boolean isExistantParId(Long id) throws ServiceException {
+	public boolean isExistantParId(Long compteId) throws ServiceException {
 
-		Assert.notNull(id, () -> "L'ID pour la vérification de l'existence d'un compte de type '" + getTClass().getSimpleName() + "' est obligatoire");
+		if ( compteId == null ) {
+			throw new ServiceException(
+					ErreurProgrammation.ID_NULL,
+					getTClass().getSimpleName());
+		}
 
 		try {
-			return this.getRepository().existsById(id);
+			return this.getRepository().existsById(compteId);
 		}
 		catch (Throwable t) {
 			throw new ServiceException(
 					t,
-					CompteTechniqueErreur.EXISTENCE_PAR_ID,
-					getTClass().getSimpleName(),
-					id);
+					ErreurTechnique.TECH_EXISTANCE_COMPTE_PAR_ID,
+					compteId,
+					getTClass().getSimpleName() );
 		}
 	}
 
 	public T rechercherParIdentifiant(String identifiant) throws ServiceException {
 
-		Assert.notNull(identifiant, () -> "L'IDENTIFIANT pour la recherche d'un compte de type '" + getTClass().getSimpleName() + "' est obligatoire");
+		if ( identifiant == null || identifiant.isBlank() ) {
+			throw new ServiceException(
+					ErreurProgrammation.IDENTIFIANT_NULL,
+					getTClass().getSimpleName());
+		}
 		
 		try {
 			Optional<T> optional = getRepository().findByIdentifiant(identifiant);
@@ -60,7 +71,7 @@ public abstract class CompteService<T extends Compte> {
 		catch (Throwable t) {
 			throw new ServiceException(
 					t,
-					CompteTechniqueErreur.RECHERCHE_PAR_IDENTIFIANT_FONCTIONNEL,
+					ErreurTechnique.TECH_RECHERCHE_COMPTE_PAR_IDENTIFIANT,
 					getTClass().getSimpleName(),
 					identifiant);
 		}
@@ -68,7 +79,11 @@ public abstract class CompteService<T extends Compte> {
 
 	public boolean isExistantParIdentifiant(String identifiant) throws ServiceException {
 
-		Assert.notNull(identifiant, () -> "L'IDENTIFIANT pour la vérification de l'existence d'un compte de type '" + getTClass().getSimpleName() + "' est obligatoire");
+		if ( identifiant == null || identifiant.isBlank() ) {
+			throw new ServiceException(
+					ErreurProgrammation.IDENTIFIANT_NULL,
+					getTClass().getSimpleName());
+		}
 		
 		try {
 			return this.getRepository().existsByIdentifiant(identifiant);
@@ -76,9 +91,9 @@ public abstract class CompteService<T extends Compte> {
 		catch (Throwable t) {
 			throw new ServiceException(
 					t,
-					CompteTechniqueErreur.EXISTENCE_PAR_IDENTIFIANT_FONCTIONNEL,
+					ErreurTechnique.TECH_EXISTANCE_COMPTE_PAR_IDENTIFIANT,
 					getTClass().getSimpleName(),
-					identifiant);
+					identifiant );
 		}
 	}
 
@@ -90,14 +105,18 @@ public abstract class CompteService<T extends Compte> {
 		catch (Throwable t) {
 			throw new ServiceException (
 					t,
-					CompteTechniqueErreur.RECHERCHE_TOUS,
+					ErreurTechnique.TECH_RECHERCHE_COMPTE_TOUS,
 					getTClass().getSimpleName());
 		}
 	}
 
 	public List<T> rechercherTous(Sort tri) throws ServiceException {
 
-		Assert.notNull(tri, () -> "Le TRI pour la recherche de tous les comptes de type '" + getTClass().getSimpleName() + "' est obligatoire");
+		if ( tri == null ) {
+			throw new ServiceException(
+					ErreurProgrammation.TRI_NULL,
+					getTClass().getSimpleName());
+		}
 
 		try {
 			return getRepository().findAll(tri);
@@ -105,7 +124,7 @@ public abstract class CompteService<T extends Compte> {
 		catch (Throwable t) {
 			throw new ServiceException (
 					t,
-					CompteTechniqueErreur.RECHERCHE_TOUS,
+					ErreurTechnique.TECH_RECHERCHE_COMPTE_TOUS,
 					getTClass().getSimpleName());
 		}
 	}
@@ -118,14 +137,18 @@ public abstract class CompteService<T extends Compte> {
 		catch (Throwable t) {
 			throw new ServiceException (
 					t,
-					CompteTechniqueErreur.SUPPRESSION_TOUS,
+					ErreurTechnique.TECH_SUPPRESSION_COMPTE_TOUS,
 					getTClass().getSimpleName());
 		}
 	}
 
 	public final T creerCompte(T compte) throws ServiceException {
 
-		Assert.notNull(compte, () -> "Le COMPTE à créer de type '" + getTClass().getSimpleName() + "' est obligatoire");
+		if ( compte == null ) {
+			throw new ServiceException(
+					ErreurProgrammation.COMPTE_NULL,
+					getTClass().getSimpleName());
+		}
 
 		compte = controlerEtPreparerPourCreation(compte);
 
@@ -134,53 +157,57 @@ public abstract class CompteService<T extends Compte> {
 
 	public final T modifierCompte(T compte) throws ServiceException {
 
-		Assert.notNull(compte, () -> "Le COMPTE à modifier de type '" + getTClass().getSimpleName() + "' est obligatoire");
+		if ( compte == null ) {
+			throw new ServiceException(
+					ErreurProgrammation.COMPTE_NULL,
+					getTClass().getSimpleName());
+		}
 
 		compte = controlerEtPreparerPourModification(compte);
 
 		return enregistrer(compte);
 	}
 
-	public final void supprimerCompte(T compte) throws ServiceException {
+	public final void supprimerCompte(Long compteId) throws ServiceException {
 
-		Assert.notNull(compte, () -> "Le COMPTE à supprimer de type '" + getTClass().getSimpleName() + "' est obligatoire");
+		if ( compteId == null ) {
+			throw new ServiceException(
+					ErreurProgrammation.ID_NULL,
+					getTClass().getSimpleName());
+		}
 
-		compte = controlerEtPreparerPourSuppression(compte);
+		T compte = controlerEtPreparerPourSuppression(compteId);
 
 		supprimer(compte);
 	}
 
-	protected T controlerEtPreparerPourCreation(T compte) throws ServiceException {
+	protected T enregistrer(T compte) throws ServiceException {
 
-		return compte;
-	}
-	
-	protected T controlerEtPreparerPourModification(T compte) throws ServiceException {
-
-		return compte;
-	}
-
-	protected T controlerEtPreparerPourSuppression(T compte) throws ServiceException {
-
-		verifierAbsenceOperationAssociee(compte.getId());
+		if ( compte == null ) {
+			throw new ServiceException(
+					ErreurProgrammation.COMPTE_NULL,
+					getTClass().getSimpleName());
+		}
 		
-		return compte;
-	}
-
-	private T enregistrer(T compte) throws ServiceException {
-
 		try {
 			return getRepository().save(compte);
 		}
 		catch (Throwable t) {
 			throw new ServiceException (
 					t,
-					CompteTechniqueErreur.ENREGISTREMENT,
+					ErreurTechnique.TECH_ENREGISTREMENT_COMPTE,
+					getTClass().getSimpleName(),
 					compte.getIdentifiant());
 		}
 	}
 
-	private void supprimer(T compte) throws ServiceException {
+	protected void supprimer(T compte) throws ServiceException {
+
+		if ( compte == null ) {
+			throw new ServiceException(
+					ErreurProgrammation.COMPTE_NULL,
+					getTClass().getSimpleName());
+		}
 
 		try {
 			this.getRepository().delete(compte);
@@ -188,33 +215,135 @@ public abstract class CompteService<T extends Compte> {
 		catch (Throwable t) {
 			throw new ServiceException (
 					t,
-					CompteTechniqueErreur.SUPPRESSION,
+					ErreurTechnique.TECH_SUPPRESSION_COMPTE,
+					getTClass().getSimpleName(),
 					compte.getIdentifiant());
 		}
 	}
+
+	protected T controlerEtPreparerPourCreation(T compte) throws ServiceException {
+
+		if ( compte == null ) {
+			throw new ServiceException(
+					ErreurProgrammation.COMPTE_NULL,
+					getTClass().getSimpleName());
+		}
+		
+		verifierCompteNonEnregistre(compte.getId());
+		verifierIdentifiantValideEtUnique(compte.getId(), compte.getIdentifiant());
+		
+		return compte;
+	}
 	
-	private void verifierAbsenceOperationAssociee(Long id) throws ServiceException {
+	protected T controlerEtPreparerPourModification(T compte) throws ServiceException {
+
+		if ( compte == null ) {
+			throw new ServiceException(
+					ErreurProgrammation.COMPTE_NULL,
+					getTClass().getSimpleName());
+		}
+
+		verifierCompteEnregistre(compte.getId());
+		verifierIdentifiantValideEtUnique(compte.getId(), compte.getIdentifiant());
+
+		return compte;
+	}
+
+	protected T controlerEtPreparerPourSuppression(Long compteId) throws ServiceException {
+
+		if ( compteId == null ) {
+			throw new ServiceException(
+					ErreurProgrammation.ID_NULL,
+					getTClass().getSimpleName());
+		}
+
+		verifierCompteEnregistre(compteId);
+		verifierAbsenceOperationAssociee(compteId);
+		
+		return rechercherParId(compteId);
+	}
+
+	private void verifierCompteEnregistre(Long compteId) throws ServiceException {
+
+		if ( compteId == null || ! isExistantParId(compteId) ) {
+			throw new ServiceException (
+					ErreurFonctionnelle.COMPTE_NON_ENREGISTRE_PAR_ID,
+					getTClass().getSimpleName(),
+					compteId);
+		}
+	}
+	
+	private void verifierCompteNonEnregistre(Long compteId) throws ServiceException {
+		
+		if ( compteId != null && isExistantParId(compteId) ) {
+			throw new ServiceException (
+					ErreurFonctionnelle.COMPTE_DEJA_ENREGISTRE_PAR_ID,
+					compteId,
+					getTClass().getSimpleName());
+		}
+	}
+
+	private void verifierIdentifiantValideEtUnique(
+			@Nullable Long compteId,
+			String compteIdentifiant) 
+					throws ServiceException {
+
+		if ( compteIdentifiant == null || compteIdentifiant.isBlank() ) {
+			throw new ServiceException (
+					ErreurFonctionnelle.COMPTE_IDENTIFIANT_INVALIDE,
+					getTClass().getSimpleName());
+		}
+
+		boolean isIdentifiantCreeOuModifie;
+		boolean isIdentifiantDejaUtilise;
+		
+		if ( compteId == null ) {
+			isIdentifiantCreeOuModifie = true;
+			isIdentifiantDejaUtilise = isExistantParIdentifiant(compteIdentifiant);
+		}
+		else {
+			try {
+				isIdentifiantCreeOuModifie = ! getRepository().existsByIdentifiantAndId(compteIdentifiant, compteId);
+				isIdentifiantDejaUtilise = getRepository().existsByIdentifiantAndIdNot(compteIdentifiant, compteId); 
+			}
+			catch ( Throwable t ) {
+				throw new ServiceException (
+						t,
+						ErreurTechnique.TECH_EXISTANCE_COMPTE_PAR_IDENTIFIANT,
+						compteIdentifiant);
+			}
+		}
+		
+		if ( isIdentifiantCreeOuModifie && isIdentifiantDejaUtilise )
+		{
+			throw new ServiceException (
+					ErreurFonctionnelle.COMPTE_IDENTIFIANT_DEJA_UTILISE,
+					getTClass().getSimpleName(),
+					compteIdentifiant);
+		}
+	}
+	
+	private void verifierAbsenceOperationAssociee(Long compteId) throws ServiceException {
 	
 		int nombreOperationAssociee;
 		try {
-			nombreOperationAssociee = getRepository().compterOperationByCompteId(id);
+			nombreOperationAssociee = getRepository().countOperationByCompteId(compteId);
 		}
 		catch (Throwable t) {
 			throw new ServiceException (
 					t,
-					CompteTechniqueErreur.COMPTAGE_USAGE_OPERATION_PAR_ID,
-					getTClass().getSimpleName(),
-					id);
+					ErreurTechnique.TECH_RECHERCHE_NOMBRE_OPERATION_PAR_COMPTE_ID,
+					compteId);
 		}
 		
 		if ( nombreOperationAssociee > 0 ) {
 			throw new ServiceException (
-					CompteFonctionnelleErreur.SUPPRESSION_COMPTE_AVEC_OPERATION,
-					id,
+					ErreurFonctionnelle.COMPTE_SUPPRESSION_AVEC_OPERATION,
+					compteId,
 					nombreOperationAssociee);
 		}
 	}
 	
-	public abstract Class<T> getTClass();
-	public abstract CompteRepository<T> getRepository();
+	protected abstract Class<T> getTClass();
+	protected abstract CompteRepository<T> getRepository();
 }
